creating a go version of the api and add the getTree endpoint
