--- backend/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "alxcode-backend",
-  "version": "1.0.0",
-  "description": "the backend for the sandbox code editer",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js",
-    "dev": "nodemon index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AjwadG/ALXCode.git"
-  },
-  "author": "AjwadG",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/AjwadG/ALXCode/issues"
-  },
-  "homepage": "https://github.com/AjwadG/ALXCode#readme",
-  "dependencies": {
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "ejs": "^3.1.10",
-    "express": "^4.19.2",
-    "socket.io": "^4.7.5"
-  }
-}
