--- backend/public/style/index.css
@@ -1,47 +1,0 @@
-body {
-  background-color: #f5f5f5;
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-.container {
-  background-color: #fff;
-  border-radius: 5px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.3);
-  padding: 20px;
-  width: 400px;
-  height: 300px;
-  overflow-y: scroll;
-}
-.output {
-  text-align: left;
-  background-color: rgb(21, 24, 45);
-  font-size: 15px;
-  color: #fff;
-  padding: 5px;
-}
-.input {
-  margin-top: 5px;
-}
-input[type="text"] {
-  width: 250px;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-button {
-  background-color: #4c5baf;
-  color: #fff;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-button:hover {
-  background-color: #45a049;
-}
