--- backend/views/index.ejs
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>home page</title>
-    <script src="https://cdn.socket.io/socket.io-3.0.1.min.js"></script>
-    <script type="module" src="scripts/index.js"></script>
-    <link rel="stylesheet" href="style/index.css" />
-  </head>
-  <body>
-    <div class="container"></div>
-    <div class="input">
-      <input class="comand" type="text" /> <button class="send">send</button>
-    </div>
-  </body>
-</html>
